+++ tests/opts.cpp
@@ -24,55 +24,62 @@ TEST(opts_test, did_not_register_next) {
 TEST(opts_test, invalid_opts_empty_name_1) {
 	auto argv = get_argv({"./executable", "-", "n", "10"});
 
-	EXPECT_THROW_TGEN_PREFIX(tgen::register_gen(argv.size() - 1, argv.data()),
-							 "invalid opt");
+	EXPECT_THROW_TGEN_PREFIX(
+		tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data()),
+		"invalid opt");
 }
 
 TEST(opts_test, invalid_opts_empty_name_2) {
 	auto argv = get_argv({"./executable", "--", "n", "10"});
 
-	EXPECT_THROW_TGEN_PREFIX(tgen::register_gen(argv.size() - 1, argv.data()),
-							 "invalid opt");
+	EXPECT_THROW_TGEN_PREFIX(
+		tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data()),
+		"invalid opt");
 }
 
 TEST(opts_test, invalid_opts_empty_key_before_eq) {
 	auto argv = get_argv({"./executable", "-=10"});
 
-	EXPECT_THROW_TGEN_PREFIX(tgen::register_gen(argv.size() - 1, argv.data()),
-							 "expected non-empty key/value in opt");
+	EXPECT_THROW_TGEN_PREFIX(
+		tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data()),
+		"expected non-empty key/value in opt");
 }
 
 TEST(opts_test, invalid_opts_empty_value_after_eq) {
 	auto argv = get_argv({"./executable", "-n="});
 
-	EXPECT_THROW_TGEN_PREFIX(tgen::register_gen(argv.size() - 1, argv.data()),
-							 "expected non-empty key/value in opt");
+	EXPECT_THROW_TGEN_PREFIX(
+		tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data()),
+		"expected non-empty key/value in opt");
 }
 
 TEST(opts_test, invalid_opts_empty_value_after_space) {
 	auto argv = get_argv({"./executable", "-n"});
 
-	EXPECT_THROW_TGEN_PREFIX(tgen::register_gen(argv.size() - 1, argv.data()),
-							 "value cannot be empty");
+	EXPECT_THROW_TGEN_PREFIX(
+		tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data()),
+		"value cannot be empty");
 }
 
 TEST(opts_test, invalid_opts_repeated_key_equal) {
 	auto argv = get_argv({"./executable", "-n", "10", "-n=20"});
 
-	EXPECT_THROW_TGEN_PREFIX(tgen::register_gen(argv.size() - 1, argv.data()),
-							 "cannot have repeated keys");
+	EXPECT_THROW_TGEN_PREFIX(
+		tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data()),
+		"cannot have repeated keys");
 }
 
 TEST(opts_test, invalid_opts_repeated_key_space) {
 	auto argv = get_argv({"./executable", "-n", "10", "-n", "20"});
 
-	EXPECT_THROW_TGEN_PREFIX(tgen::register_gen(argv.size() - 1, argv.data()),
-							 "cannot have repeated keys");
+	EXPECT_THROW_TGEN_PREFIX(
+		tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data()),
+		"cannot have repeated keys");
 }
 
 TEST(opts_test, has_opt_named) {
 	auto argv = get_argv({"./executable", "-n", "10"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_EQ(tgen::has_opt("n"), true);
 	EXPECT_EQ(tgen::has_opt("m"), false);
@@ -80,7 +87,7 @@ TEST(opts_test, has_opt_named) {
 
 TEST(opts_test, has_opt_named_char) {
 	auto argv = get_argv({"./executable", "-n", "10"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_EQ(tgen::has_opt('n'), true);
 	EXPECT_EQ(tgen::has_opt('m'), false);
@@ -88,7 +95,7 @@ TEST(opts_test, has_opt_named_char) {
 
 TEST(opts_test, has_opt_positional) {
 	auto argv = get_argv({"./executable", "-n", "10", "-10"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_EQ(tgen::has_opt(0), true);
 	EXPECT_EQ(tgen::has_opt(1), false);
@@ -96,14 +103,14 @@ TEST(opts_test, has_opt_positional) {
 
 TEST(opts_test, opt_named_not_found) {
 	auto argv = get_argv({"./executable", "-n", "10", "-10"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_THROW_TGEN_PREFIX(tgen::opt<int>("m"), "cannot find opt with key m");
 }
 
 TEST(opts_test, opt_named_invalid_conversion) {
 	auto argv = get_argv({"./executable", "-n", "value", "-10"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_THROW_TGEN_PREFIX(tgen::opt<int>("n"),
 							 "invalid value `value` for type i");
@@ -111,7 +118,7 @@ TEST(opts_test, opt_named_invalid_conversion) {
 
 TEST(opts_test, opt_named_invalid_conversion_bool) {
 	auto argv = get_argv({"./executable", "-b", "tru"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_THROW_TGEN_PREFIX(tgen::opt<bool>("b"),
 							 "invalid value `tru` for type b");
@@ -119,7 +126,7 @@ TEST(opts_test, opt_named_invalid_conversion_bool) {
 
 TEST(opts_test, opt_named) {
 	auto argv = get_argv({"./executable", "-n", "10", "-10", "-m", "true"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_EQ(tgen::opt<int>("n"), 10);
 	EXPECT_EQ(tgen::opt<bool>("m"), true);
@@ -127,7 +134,7 @@ TEST(opts_test, opt_named) {
 
 TEST(opts_test, opt_named_char) {
 	auto argv = get_argv({"./executable", "-n", "10", "-10", "-m", "true"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_EQ(tgen::opt<int>('n'), 10);
 	EXPECT_EQ(tgen::opt<bool>('m'), true);
@@ -135,21 +142,21 @@ TEST(opts_test, opt_named_char) {
 
 TEST(opts_test, opt_named_default) {
 	auto argv = get_argv({"./executable", "-n", "10", "-10"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_EQ(tgen::opt<int>("m", 20), 20);
 }
 
 TEST(opts_test, opt_positional_not_found) {
 	auto argv = get_argv({"./executable", "-n", "10", "-10"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_THROW_TGEN_PREFIX(tgen::opt<int>(1), "cannot find opt at index 1");
 }
 
 TEST(opts_test, opt_positional) {
 	auto argv = get_argv({"./executable", "-n", "10", "-10"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_EQ(tgen::opt<int>(0), -10);
 }
@@ -193,7 +200,7 @@ TEST(opts_test, set_cpp_version) {
 
 TEST(opts_test, set_cpp_version_opt) {
 	auto argv = get_argv({"./executable", "tgen::CPP:20"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 
 	EXPECT_EQ(tgen::detail::cpp.version_, 20);
 }
@@ -234,37 +241,37 @@ TEST(opts_test, set_compiler) {
 
 TEST(opts_test, set_compiler_opt) {
 	auto argv = get_argv({"./executable", "tgen::GCC"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 	EXPECT_EQ(tgen::detail::compiler.kind_, tgen::gcc().kind_);
 	EXPECT_EQ(tgen::detail::compiler.major_, tgen::gcc().major_);
 	EXPECT_EQ(tgen::detail::compiler.minor_, tgen::gcc().minor_);
 
 	argv = get_argv({"./executable", "tgen::CLANG"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 	EXPECT_EQ(tgen::detail::compiler.kind_, tgen::clang().kind_);
 	EXPECT_EQ(tgen::detail::compiler.major_, tgen::clang().major_);
 	EXPECT_EQ(tgen::detail::compiler.minor_, tgen::clang().minor_);
 
 	argv = get_argv({"./executable", "tgen::GCC:17"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 	EXPECT_EQ(tgen::detail::compiler.kind_, tgen::gcc(17).kind_);
 	EXPECT_EQ(tgen::detail::compiler.major_, tgen::gcc(17).major_);
 	EXPECT_EQ(tgen::detail::compiler.minor_, tgen::gcc(17).minor_);
 
 	argv = get_argv({"./executable", "tgen::CLANG:17"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 	EXPECT_EQ(tgen::detail::compiler.kind_, tgen::clang(17).kind_);
 	EXPECT_EQ(tgen::detail::compiler.major_, tgen::clang(17).major_);
 	EXPECT_EQ(tgen::detail::compiler.minor_, tgen::clang(17).minor_);
 
 	argv = get_argv({"./executable", "tgen::GCC:17.2"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 	EXPECT_EQ(tgen::detail::compiler.kind_, tgen::gcc(17, 2).kind_);
 	EXPECT_EQ(tgen::detail::compiler.major_, tgen::gcc(17, 2).major_);
 	EXPECT_EQ(tgen::detail::compiler.minor_, tgen::gcc(17, 2).minor_);
 
 	argv = get_argv({"./executable", "tgen::CLANG:17.2"});
-	tgen::register_gen(argv.size() - 1, argv.data());
+	tgen::register_gen(static_cast<int>(argv.size()) - 1, argv.data());
 	EXPECT_EQ(tgen::detail::compiler.kind_, tgen::clang(17, 2).kind_);
 	EXPECT_EQ(tgen::detail::compiler.major_, tgen::clang(17, 2).major_);
 	EXPECT_EQ(tgen::detail::compiler.minor_, tgen::clang(17, 2).minor_);
